+++ package.json
@@ -35,8 +35,8 @@
         "react-dom": "16.8.4",
         "react-foco": "1.2.0",
         "react-hot-loader": "4.8.0",
-        "react-router": "4.3.1",
-        "react-router-dom": "4.3.1",
+        "react-router": "5.1.2",
+        "react-router-dom": "5.1.2",
         "react-ui-validations": "0.3.0",
         "remarkable": "1.7.1",
         "retail-ui": "0.44.1"
