+++ src/Containers/TriggerAddContainer.jsx
@@ -1,6 +1,6 @@
 // @flow
 import * as React from "react";
-import type { ContextRouter } from "react-router-dom";
+import type { ContextRouter, useHistory } from "react-router-dom";
 import { ValidationContainer } from "react-ui-validations";
 import Button from "retail-ui/components/Button";
 import type { IMoiraApi } from "../Api/MoiraApi";
@@ -129,7 +129,7 @@ class TriggerAddContainer extends React.Component<Props, State> {
 
     async handleSubmit() {
         let { trigger } = this.state;
-        const { history, moiraApi } = this.props;
+        const { moiraApi } = this.props;
         // ToDo отказаться от вереницы if
         if (this.validationContainer.current !== null) {
             const isValid: boolean = await this.validationContainer.current.validate();
@@ -150,6 +150,7 @@ class TriggerAddContainer extends React.Component<Props, State> {
                 }
                 try {
                     const { id } = await moiraApi.addTrigger(trigger);
+                    const history = useHistory();
                     history.push(getPageLink("trigger", id));
                 } catch (error) {
                     this.setState({ error: error.message, loading: false });
