+++ src/Containers/TriggerDuplicateContainer.jsx
@@ -1,6 +1,6 @@
 // @flow
 import * as React from "react";
-import type { ContextRouter } from "react-router-dom";
+import type { ContextRouter, useHistory } from "react-router-dom";
 import { ValidationContainer } from "react-ui-validations";
 import Button from "retail-ui/components/Button";
 import type { IMoiraApi } from "../Api/MoiraApi";
@@ -119,7 +119,7 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
 
     async handleSubmit() {
         let { trigger } = this.state;
-        const { history, moiraApi } = this.props;
+        const { moiraApi } = this.props;
         const isValid = await this.validateForm();
         if (isValid && trigger) {
             this.setState({ loading: true });
@@ -138,6 +138,7 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
             }
             try {
                 const { id } = await moiraApi.addTrigger(trigger);
+                const history = useHistory()
                 history.push(getPageLink("trigger", id));
             } catch (error) {
                 this.setState({ error: error.message, loading: false });
