+++ src/pages/trigger/trigger.jsx
@@ -1,7 +1,7 @@
 // @flow
 import * as React from "react";
 import queryString from "query-string";
-import type { ContextRouter } from "react-router-dom";
+import type { ContextRouter, useHistory } from "react-router-dom";
 import isEqual from "lodash/isEqual";
 import type { Trigger, TriggerState } from "../../Domain/Trigger";
 import type { Event } from "../../Domain/Event";
@@ -10,6 +10,7 @@ import type { Maintenance } from "../../Domain/Maintenance";
 import { withMoiraApi } from "../../Api/MoiraApiInjection";
 import { setMetricMaintenance, setTriggerMaintenance } from "../../Domain/Maintenance";
 import transformPageFromHumanToProgrammer from "../../logic/transformPageFromHumanToProgrammer";
+
 
 type Props = ContextRouter & { moiraApi: IMoiraApi };
 
@@ -181,9 +182,9 @@ class TriggerPage extends React.Component<Props, State> {
     };
 
     changeLocationSearch = update => {
-        const { history, location } = this.props;
+        const { location } = this.props;
         const locationSearch = TriggerPage.parseLocationSearch(location.search);
-
+        const history = useHistory();
         history.push(
             `?${queryString.stringify(
                 { ...locationSearch, ...update },
